collector_telegram/requirements.txt
telethon
httpx
pydantic
python-dotenv
cryptography
